pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sar-speckle"
version = "0.1.0"
authors = [{ name = "Félix", email = "felixmaral131@gmail.com" }]
description = "Research toolkit for SAR speckle simulation and mitigation"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "Pillow>=9.0",
  "matplotlib>=3.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
speckle = "sar_speckle.cli:main"

[tool.ruff]
line-length = 100